% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{MIApassing}
\alias{MIApassing}
\title{Miami Dolphins passing data}
\format{A data frame of 16 rows and 4 columns, with each row being a regular season game
\describe{
 \item{comp}{Numeric values showing the completed passes}
 \item{att}{Numeric values showing the passes attempted}
 \item{yards}{Numeric values showing the total yards gained by passing}
 \item{season}{Charakter values showing the year of the season}
}}
\source{
https://www.pro-football-reference.com/teams/mia/2017/gamelog/ opened on 30 Dezember 2018
}
\usage{
MIApassing
}
\description{
A dataset containing data from the 2017 season of the NFL team Miami Dolphins.
}
\keyword{datasets}
