% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_prediction.R
\name{SLR_prediction.data.frame}
\alias{SLR_prediction.data.frame}
\title{SLR prediction S3 method for data frames}
\usage{
\method{SLR_prediction}{data.frame}(y, col_response = 1,
  col_exploratory = 2, pre_x_value = mean(y[, col_exploratory]))
}
\arguments{
\item{y}{a numeric \code{data.frame} containing the response and exploratory variable}

\item{col_response}{column number of the response variable within y (default is 1)}

\item{col_exploratory}{column number of the exploratory variable within y (default is 2)}

\item{pre_x_value}{the numeric value (within the range of the exploratory vector) for which the prediction is made}
}
\value{
the predicted value of the response variable
}
\description{
Predicting with a simple linear regression model
}
\details{
The function first builds a simple linear regression model with one response and one exploratory variable. Then the prediction for a given value of the exploratory variable (default the mean and no out of sample predictions) gets returned.
}
\examples{
special <- 6
Y<-data.frame(c(1:5),c(1,3,4.5,7,8))
SLR_prediction(Y,pre_x_value = special)
}
\seealso{
\link{SLR_prediction_vis}
}
\author{
Emanuel Sommer
}
