% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_prediction.R
\name{SLR_prediction.default}
\alias{SLR_prediction.default}
\title{SLR prediction default method}
\usage{
\method{SLR_prediction}{default}(y)
}
\arguments{
\item{y}{an object of a wrong class}
}
\description{
SLR prediction default method
}
\author{
Emanuel Sommer
}
