% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_prediction.R
\name{SLR_prediction.numeric}
\alias{SLR_prediction.numeric}
\title{SLR prediction numeric S3 method}
\usage{
\method{SLR_prediction}{numeric}(y, exploratory,
  pre_x_value = mean(exploratory))
}
\arguments{
\item{y}{a numeric vector (the response variable)}

\item{exploratory}{a numeric vector (the exploratory variable with the same length as y)}

\item{pre_x_value}{the numeric value (within the range of the exploratory vector) for which the prediction is made}
}
\value{
the predicted value of the response variable
}
\description{
Predicting with a simple linear regression model
}
\details{
The function first builds a simple linear regression model with one response and one exploratory variable. Then the prediction for a given value of the exploratory variable (default the mean and no out of sample predictions) gets returned.
}
\examples{
##numeric S3 method
response <- c(1:5)
exploratory <- c(1,3,4.5,7,8)
special <- 6
SLR_prediction(response,exploratory,pre_x_value = special)
}
\seealso{
\code{\link{SLR_prediction_vis}}
}
\author{
Emanuel Sommer
}
