% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_prediction.R
\name{SLR_prediction}
\alias{SLR_prediction}
\title{SLR prediction}
\usage{
SLR_prediction(y, exploratory, col_response, col_exploratory, pre_x_value)
}
\arguments{
\item{y}{depends on the class of the input: for
\describe{
 \item{numeric vectors}{a numeric vector representing the response variable}
 \item{data frames}{a numeric \code{data.frame} containing the response and exploratory variable}
 \item{matrix}{a numeric \code{matrix} containing the response and exploratory variable}
 }}

\item{exploratory}{numeric vector representing the exploratory variable, must have the same length as y (for the input class numeric only)}

\item{col_response}{column number of the response variable within y (default is 1) (for data frames and matrices only)}

\item{col_exploratory}{column number of the exploratory variable within y (default is 2) (for data frames and matrices only)}

\item{pre_x_value}{the numeric value (within the range of the exploratory vector) for which the prediction is made}
}
\value{
the predicted value of the response variable
}
\description{
Predicting with a simple linear regression model. (S3 generic function)
}
\details{
The generic function first builds a simple linear regression model with one response and one exploratory variable. Then the prediction for a given value of the exploratory variable (default the mean and no out of sample predictions) gets returned.
}
\examples{
##numeric S3 method
response <- c(1:5)
exploratory <- c(1,3,4.5,7,8)
special <- 6
SLR_prediction(response,exploratory,pre_x_value = special)

##data.frame S3 method
Y<-data.frame(c(1:5),c(1,3,4.5,7,8))
SLR_prediction(Y,pre_x_value = special)

##matrix S3 method is eqivalently used as the one for data frames
}
\seealso{
\code{\link{SLR_prediction_vis}} \code{\link{SLR_prediction.numeric}} \code{\link{SLR_prediction.data.frame}} \code{\link{SLR_prediction.matrix}}
}
\author{
Emanuel Sommer
}
