% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_prediction_vis.R
\name{SLR_prediction_vis.default}
\alias{SLR_prediction_vis.default}
\title{SLR prediction visualization default method}
\usage{
\method{SLR_prediction_vis}{default}(y)
}
\arguments{
\item{y}{an object of a wrong class}
}
\description{
SLR prediction visualization default method
}
\author{
Emanuel Sommer
}
