% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_prediction_vis.R
\name{SLR_prediction_vis.numeric}
\alias{SLR_prediction_vis.numeric}
\title{SLR prediction visualization numeric S3 method}
\usage{
\method{SLR_prediction_vis}{numeric}(y, exploratory,
  pre_x_value = mean(exploratory), pre_col = "red", reg_col = "blue")
}
\arguments{
\item{y}{a numeric vector (the response variable)}

\item{exploratory}{a numeric vector (the exploratory variable with the same length as y)}

\item{pre_x_value}{the numeric value (within the range of the exploratory vector) for which the prediction is made}

\item{pre_col}{color of the highlighted prediction within the scatterplot}

\item{reg_col}{color of the regression line}
}
\value{
a scatterplot of the two variables + regression line + highlighted prediction
}
\description{
Simple visualization tool for predictions within a simple linear regression model.
}
\details{
The generic function first builds a simple linear regression model with one response and one exploratory variable. Then the prediction for a given value of the exploratory variable (default the mean and no out of sample predictions) gets computed.(see \link{SLR_prediction}) Finally using these bits of information a scatterplot of the two variables with the regression line gets returned. Moreover the predicition will be highlighted in the scatterplot and the exact values of the predicition will be displayed above the plot.
}
\examples{
response <- c(1:5)
exploratory <- c(1,3,4.5,7,8)
special <- 6
SLR_prediction_vis(response,exploratory,pre_x_value = special)
}
\seealso{
\link{SLR_prediction}
}
\author{
Emanuel Sommer
}
