% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLR_prediction.R
\name{SLR_prediction}
\alias{SLR_prediction}
\title{SLR prediction}
\usage{
SLR_prediction(y, ...)
}
\arguments{
\item{y}{depends on the class of the input: for
\describe{
 \item{numeric vectors}{see \link{SLR_prediction.numeric}}
 \item{data frames}{see \link{SLR_prediction.data.frame}}
 \item{matrix}{see \link{SLR_prediction.matrix}}
 }}

\item{...}{also depending on the class of the desired input}
}
\value{
the predicted value of the response variable
}
\description{
Predicting with a simple linear regression model. (S3 generic function)
}
\details{
The generic function first builds a simple linear regression model with one response and one exploratory variable. Then the prediction for a given value of the exploratory variable (default the mean and no out of sample predictions) gets returned.
}
\examples{
##numeric S3 method
response <- c(1:5)
exploratory <- c(1,3,4.5,7,8)
special <- 6
SLR_prediction(response,exploratory,pre_x_value = special)

##data.frame S3 method
Y<-data.frame(c(1:5),c(1,3,4.5,7,8))
SLR_prediction(Y,pre_x_value = special)

##matrix S3 method is eqivalently used as the one for data frames
}
\seealso{
\link{SLR_prediction_vis}
}
\author{
Emanuel Sommer
}
